Remove EDH donation totals api methods & tests
