+++ source/api/donation-totals/__tests__/totals-test.js
@@ -1,75 +1,25 @@
-import moxios from 'moxios'
-import { instance, servicesAPI, updateClient } from '../../../utils/client'
-import { fetchDonationTotals } from '..'
+import { instance, servicesAPI } from '../../../utils/client'
+import { fetchDonationTotals, deserializeDonationTotals } from '..'
 
-import {
-  fetchDonationTotals as fetchEDHDonationTotals,
-  deserializeDonationTotals as deserializeEDHTotals
-} from '../everydayhero'
-
-import {
-  fetchDonationTotals as fetchJGDonationTotals,
-  deserializeDonationTotals as deserializeJGTotals
-} from '../justgiving'
-
-describe('Fetch Donation Totals', () => {
-  it('throws if no params are passed in', () => {
-    const test = () => fetchDonationTotals()
-    expect(test).to.throw
-  })
-
-  describe('Fetch EDH Donation Totals', () => {
+describe('Donation Totals', () => {
+  describe('Fetch Donation Totals', () => {
     beforeEach(() => {
-      moxios.install(instance)
-    })
-
-    afterEach(() => {
-      moxios.uninstall(instance)
-    })
-
-    it('uses the correct url to fetch totals for a campaign', done => {
-      fetchEDHDonationTotals({ campaign_id: 'au-6839' })
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/search/totals'
-        )
-        expect(request.url).to.contain('campaign_id=au-6839')
-        done()
-      })
-    })
-
-    it('uses the correct url to fetch totals for a charity', done => {
-      fetchEDHDonationTotals({ charity_id: 'au-28' })
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/search/totals'
-        )
-        expect(request.url).to.contain('charity_id=au-28')
-        done()
-      })
-    })
-  })
-
-  describe('Fetch JG Donation Totals', () => {
-    beforeEach(() => {
-      updateClient({
-        baseURL: 'https://api.justgiving.com',
-        headers: { 'x-api-key': 'abcd1234' }
-      })
       moxios.install(instance)
       moxios.install(servicesAPI)
     })
 
     afterEach(() => {
-      updateClient({ baseURL: 'https://everydayhero.com' })
       moxios.uninstall(instance)
       moxios.uninstall(servicesAPI)
     })
 
+    it('throws if no params are passed in', () => {
+      const test = () => fetchDonationTotals()
+      expect(test).to.throw
+    })
+
     it('uses the correct url to fetch totals for an event', done => {
-      fetchJGDonationTotals({ event: 12345 })
+      fetchDonationTotals({ event: 12345 })
       moxios.wait(() => {
         const firstRequest = moxios.requests.first()
         const secondRequest = moxios.requests.mostRecent()
@@ -86,7 +36,7 @@ describe('Fetch Donation Totals', () => {
     })
 
     it('allows the country (and currency) to be specified', done => {
-      fetchJGDonationTotals({ event: 12345, country: 'ie' })
+      fetchDonationTotals({ event: 12345, country: 'ie' })
       moxios.wait(() => {
         const request = moxios.requests.mostRecent()
         expect(request.url).to.equal(
@@ -97,7 +47,7 @@ describe('Fetch Donation Totals', () => {
     })
 
     it('uses the correct urls to fetch totals for a campaign (with offline amounts)', done => {
-      fetchJGDonationTotals({ campaign: 'my-campaign', includeOffline: true })
+      fetchDonationTotals({ campaign: 'my-campaign', includeOffline: true })
       moxios.wait(() => {
         const firstRequest = moxios.requests.at(0)
         const secondRequest = moxios.requests.at(1)
@@ -113,7 +63,7 @@ describe('Fetch Donation Totals', () => {
     })
 
     it('uses the correct url to fetch totals for multiple campaigns', done => {
-      fetchJGDonationTotals({ campaign: ['1234', '5678'] })
+      fetchDonationTotals({ campaign: ['1234', '5678'] })
       moxios.wait(() => {
         const request = moxios.requests.mostRecent()
         expect(request.url).to.equal(
@@ -124,7 +74,7 @@ describe('Fetch Donation Totals', () => {
     })
 
     it('uses the correct url to fetch totals for a charity', done => {
-      fetchJGDonationTotals({ charity: 1234 })
+      fetchDonationTotals({ charity: 1234 })
       moxios.wait(() => {
         const request = moxios.requests.mostRecent()
         expect(request.url).to.include(
@@ -134,28 +84,10 @@ describe('Fetch Donation Totals', () => {
       })
     })
   })
-})
 
-describe('Deserialize donation totals', () => {
-  describe('Deserialize EDH donation totals', () => {
+  describe('Deserialize donation totals', () => {
     it('Defaults falsy donation sums to 0', () => {
-      const deserializedDonationTotals = deserializeEDHTotals({
-        total_amount_cents: {
-          sum: null,
-          count: 0
-        }
-      })
-
-      expect(deserializedDonationTotals).to.deep.equal({
-        raised: 0,
-        donations: 0
-      })
-    })
-  })
-
-  describe('Deserialize JG donation totals', () => {
-    it('Defaults falsy donation sums to 0', () => {
-      const deserializedDonationTotals = deserializeJGTotals({
+      const deserializedDonationTotals = deserializeDonationTotals({
         raisedAmount: 0
       })
